Added negative test for the signup
